Create Homework 0
Homework 0 - Composing Functions
1. (20 Points)  Evaluate each of the following expressions. Show every step in the 
evaluation. Provide your answers as comments in the Definitions window. For 
example, to evaluate the expression 
(* (+ 4 3) 9)
 
You should provide the following comments in the Definitions window: 
;; (*(+ 4 3) 9)
;; (* 7 9)
;; 63
  
(a.) (* (- (* 5 3) (- 8 (string-length (string-append "CS" "1101")))) (/ 64 16))
 
(b.) (if (< (triple 2) (triple (/ 12 4))) 
    (sqrt (+ 4 (* 15 4)))
    107)
 
where triple is defined as 
(define (triple n)
   (* n 3))

Composing Functions  
 Provide a signature and purpose for each function you write. 
 
2. (20 Points)  In class we played with some built-in functions in the image 
library. Add the following line to the Definitions window: 
(require 2htdp/image)
 
Now click the Run button. The functions in the image library are now available to 
your program. From the Help menu, choose Help Desk. A browser window will 
open. Under the Teaching heading, choose How to Design ProgramsTeachpacks. 
In the left margin, click on HtDP/2e Teachpacks. Here you will find a description 
of all the pre-defined functions in the image.ss library. Notice the ...search 
manuals... box at the top of the page. Type in image-width to find information 
about that function. When the DrRacket Helpdesk provides the matches for 
image-width, choose the one from "2htdp/image". 
 Develop a function called four-square that consumes an image and produces an 
image that consists of 4 copies of the original image, as shown in this example: 
if the original image looks like this:
then the produced image should look like this:
Use the DrRacket help desk to learn about the following functions (available in 
the 2htdp/image library) that you might consider using for this problem: 
flip-vertical 
flip-horizontal 
beside 
above 
rotate
(You're not restricted to using these functions. You may solve the problem 
any way you want as long as the image is duplicated in the same manner 
as the image in the illustration.) You may use the image in this problem 
for testing purposes, or you may test your program with an image of your 
own choosing (see Google images). Use a square-shaped image. To copy 
an image, right-click on the image, select Copy, then in the DrRacket 
Definitions window, select Paste. You should give your image a name 
using define. 

3. (70 Points)  A local performing arts center books various events. Every event 
costs the center $2500 (for utility costs, custodial help, etc.) plus $3 per attendee 
(for program printing, etc.). You are to define a set of functions that will calculate 
the profit for an event based on the number of attendees and the price of a ticket.
Develop a function cost that consumes the number of attendees at an event and produces the cost to the performing arts center to hold the event.
Develop a function revenue that consumes the number of attendees and the price per ticket and produces the revenue generated by ticket sales.
Develop a function profit that consumes the number of attendees and the price per ticket and produces the profit for the event.
 Use defined constants where appropriate.   Make sure each function you define is 
documented with a signature and a purpose. 

4. (40 Points)  Develop a function total-house-cost. The function consumes the 
purchase price of a house and the amount of the down-payment, and produces the 
grand total paid by the buyer for the house, as described in the following 
paragraph. The following terms apply: 
The principal is the difference between the purchase price and the down-payment. 
Only the principal is used in the calculation of the interest. 
The interest is calculated using the formula for simple interest (rate * amount * time). 
The period of the loan is 20 years, and the interest rate is 3% per year. 
Bank fees (closing costs) total $1500. 
The total cost of the house is the purchase price plus the interest paid plus the fees. 
You should develop helper functions and define constants as necessary. 
Make sure each function you define is documented with a signature and a 
purpose.